backend/resume_extractor/requirements.txt
pdfplumber==0.11.0
pymongo
python-dotenv
sentence-transformers
numpy
